+++ Project1/pqueue.h
@@ -12,4 +12,4 @@ void pop_queue(pNode** head);
 void push_queue(pNode** head, void* val, int priority);
 void push_queue_end(pNode** head, void* val);
 bool is_queue_empty(pNode* head);
-int size_queue(pNode* head);
+size_t size_queue(pNode* head);
